pyproject.toml
[tool.poetry]
name = "cyberpower-pdu"
version = "0.1.0"
description = "Python API for interacting with a CyberPower PDU over the network"
authors = ["Blake Mitchell <bmitc@mit.edu>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
pysnmp-lextudio = "5.0.34"
telnetlib3 = "^2.0.4"
pyside6 = "^6.7.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.8.0"
ruff = "^0.6.5"
isort = "^5.13.2"
mypy = "^1.11.2"
poethepoet = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
extend-exclude = '''
  \.mypy_cache
| __pycache__
| \.pytest_cache
| \.venv
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
extend_skip = [".mypy_cache", "__pycache__", ".pytest_cache", ".venv"]
import_heading_future = "The `__future__` import must be listed first. Otherwise, a `SyntaxError` is emitted."
import_heading_stdlib = "Core dependencies"
import_heading_thirdparty = "Package dependencies"
import_heading_firstparty = "Project dependencies"
import_heading_localfolder = "Local dependencies"
force_sort_within_sections = true
combine_as_imports = true

[tool.mypy]
exclude = [".*/?\\.venv/.*", "/site-packages/"]
strict = true
follow_imports = "silent"

[tool.poe.tasks]
all = ["format", "typecheck", "test"]
check-imports = "isort . --check"
check-code-formatting = "black . --check"
check-formatting = ["check-code-formatting", "check-imports"]
format-imports = "isort ."
format-code = "black ."
format = ["format-code", "format-imports"]
test = "pytest"
typecheck = "mypy --package cyberpower_pdu --no-site-packages --enable-incomplete-feature=NewGenericSyntax"
